more

+++ 2_download_and_transcribe.py
@@ -4,6 +4,9 @@ import json
 import requests
 import unicodedata
 
+model = "medium"
+weights = "ggml-tiny.en.bin"
+
 def run(p):
     subprocess.run(p.split(" "))
 
@@ -12,25 +15,27 @@ eps = json.load(open("entitled-opinions.json", "r"))
 for ts, ep in eps.items():
     t = unicodedata.normalize("NFKD", ep["title"])
 
-    txf = f"./episodes/{ts}/transcript-tiny"
+    txf = f"./episodes/{ts}/transcript-{model}"
     if os.path.exists(txf + ".vtt"):
         print("Skipping", t)
         continue
 
     os.makedirs(f"./episodes/{ts}", exist_ok=True)
 
-    print("Requesting", t)
-    audio = requests.get(ep["url"])
-
-    print("Writing", t)
     src = f'./episodes/{ts}/audio.mp3'
-    with open(src, 'wb') as f:
-        f.write(audio.content)
+    if not os.path.exists(src):
+        print("Requesting", t)
+        audio = requests.get(ep["url"])
 
-    print("Re-encoding", t)
+        print("Writing", t)
+        with open(src, 'wb') as f:
+            f.write(audio.content)
+
     tgt = src.replace('mp3', 'wav')
-    run(f"ffmpeg -y -i {src} -acodec pcm_s16le -ac 1 -ar 16000 {tgt}")
+    if not os.path.exists(tgt):
+        print("Re-encoding", t)
+        run(f"ffmpeg -y -i {src} -acodec pcm_s16le -ac 1 -ar 16000 {tgt}")
 
     print("Transcribing", t)
     wcpp = "../whisper.cpp"
-    run(f"{wcpp}/main -f {tgt} --model {wcpp}/models/ggml-tiny.en.bin --output-vtt --output-file {txf}")
+    run(f"{wcpp}/main -f {tgt} --model {wcpp}/models/ggml-{model}.en.bin --output-vtt --output-file {txf}")
